+++ pkg/apis/core/v1/helper/qos/qos_test.go
@@ -39,21 +39,8 @@ func TestGetPodQOS(t *testing.T) {
 			expected: v1.PodQOSGuaranteed,
 		},
 		{
-			pod: newPod("guaranteed-with-gpu", []v1.Container{
-				newContainer("guaranteed", getResourceList("100m", "100Mi"), addResource("nvidia-gpu", "2", getResourceList("100m", "100Mi"))),
-			}),
-			expected: v1.PodQOSGuaranteed,
-		},
-		{
 			pod: newPod("guaranteed-guaranteed", []v1.Container{
 				newContainer("guaranteed", getResourceList("100m", "100Mi"), getResourceList("100m", "100Mi")),
-				newContainer("guaranteed", getResourceList("100m", "100Mi"), getResourceList("100m", "100Mi")),
-			}),
-			expected: v1.PodQOSGuaranteed,
-		},
-		{
-			pod: newPod("guaranteed-guaranteed-with-gpu", []v1.Container{
-				newContainer("guaranteed", getResourceList("100m", "100Mi"), addResource("nvidia-gpu", "2", getResourceList("100m", "100Mi"))),
 				newContainer("guaranteed", getResourceList("100m", "100Mi"), getResourceList("100m", "100Mi")),
 			}),
 			expected: v1.PodQOSGuaranteed,
@@ -72,28 +59,15 @@ func TestGetPodQOS(t *testing.T) {
 			expected: v1.PodQOSBestEffort,
 		},
 		{
-			pod: newPod("best-effort-best-effort-with-gpu", []v1.Container{
-				newContainer("best-effort", getResourceList("", ""), addResource("nvidia-gpu", "2", getResourceList("", ""))),
-				newContainer("best-effort", getResourceList("", ""), getResourceList("", "")),
-			}),
-			expected: v1.PodQOSBestEffort,
-		},
-		{
-			pod: newPod("best-effort-with-gpu", []v1.Container{
-				newContainer("best-effort", getResourceList("", ""), addResource("nvidia-gpu", "2", getResourceList("", ""))),
-			}),
-			expected: v1.PodQOSBestEffort,
-		},
-		{
 			pod: newPod("best-effort-burstable", []v1.Container{
-				newContainer("best-effort", getResourceList("", ""), addResource("nvidia-gpu", "2", getResourceList("", ""))),
+				newContainer("best-effort", getResourceList("", ""), getResourceList("", "")),
 				newContainer("burstable", getResourceList("1", ""), getResourceList("2", "")),
 			}),
 			expected: v1.PodQOSBurstable,
 		},
 		{
 			pod: newPod("best-effort-guaranteed", []v1.Container{
-				newContainer("best-effort", getResourceList("", ""), addResource("nvidia-gpu", "2", getResourceList("", ""))),
+				newContainer("best-effort", getResourceList("", ""), getResourceList("", "")),
 				newContainer("guaranteed", getResourceList("10m", "100Mi"), getResourceList("10m", "100Mi")),
 			}),
 			expected: v1.PodQOSBurstable,
@@ -132,7 +106,7 @@ func TestGetPodQOS(t *testing.T) {
 		},
 		{
 			pod: newPod("burstable-2", []v1.Container{
-				newContainer("burstable", getResourceList("0", "0"), addResource("nvidia-gpu", "2", getResourceList("100m", "200Mi"))),
+				newContainer("burstable", getResourceList("0", "0"), getResourceList("100m", "200Mi")),
 			}),
 			expected: v1.PodQOSBurstable,
 		},
