--- pkg/kubelet/gpu/BUILD
@@ -1,32 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "gpu_manager_stub.go",
-        "types.go",
-    ],
-    importpath = "k8s.io/kubernetes/pkg/kubelet/gpu",
-    deps = ["//vendor/k8s.io/api/core/v1:go_default_library"],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//pkg/kubelet/gpu/nvidia:all-srcs",
-    ],
-    tags = ["automanaged"],
-)
