--- pkg/kubelet/gpu/OWNERS
@@ -1,12 +1,0 @@
-approvers:
-- dchen1107
-- derekwaynecarr
-- vishh
-- yujuhong
-reviewers:
-- cmluciano
-- jiayingz
-- mindprince
-- RenaudWasTaken
-- vishh
-- sig-node-reviewers
