+++ test/e2e/scheduling/nvidia-gpus.go
@@ -40,53 +40,10 @@ const (
 	driverInstallTimeout = 10 * time.Minute
 )
 
-type podCreationFuncType func() *v1.Pod
-
 var (
 	gpuResourceName v1.ResourceName
 	dsYamlUrl       string
-	podCreationFunc podCreationFuncType
 )
-
-func makeCudaAdditionTestPod() *v1.Pod {
-	podName := testPodNamePrefix + string(uuid.NewUUID())
-	testPod := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: podName,
-		},
-		Spec: v1.PodSpec{
-			RestartPolicy: v1.RestartPolicyNever,
-			Containers: []v1.Container{
-				{
-					Name:  "vector-addition",
-					Image: imageutils.GetE2EImage(imageutils.CudaVectorAdd),
-					Resources: v1.ResourceRequirements{
-						Limits: v1.ResourceList{
-							gpuResourceName: *resource.NewQuantity(1, resource.DecimalSI),
-						},
-					},
-					VolumeMounts: []v1.VolumeMount{
-						{
-							Name:      "nvidia-libraries",
-							MountPath: "/usr/local/nvidia/lib64",
-						},
-					},
-				},
-			},
-			Volumes: []v1.Volume{
-				{
-					Name: "nvidia-libraries",
-					VolumeSource: v1.VolumeSource{
-						HostPath: &v1.HostPathVolumeSource{
-							Path: "/home/kubernetes/bin/nvidia/lib",
-						},
-					},
-				},
-			},
-		},
-	}
-	return testPod
-}
 
 func makeCudaAdditionDevicePluginTestPod() *v1.Pod {
 	podName := testPodNamePrefix + string(uuid.NewUUID())
@@ -163,20 +120,13 @@ func SetupNVIDIAGPUNode(f *framework.Framework, setupResourceGatherer bool) *fra
 	}
 	framework.Logf("Cluster is running on COS. Proceeding with test")
 
-	if f.BaseName == "gpus" {
-		dsYamlUrl = "https://raw.githubusercontent.com/ContainerEngine/accelerators/master/cos-nvidia-gpu-installer/daemonset.yaml"
-		gpuResourceName = v1.ResourceNvidiaGPU
-		podCreationFunc = makeCudaAdditionTestPod
+	dsYamlUrlFromEnv := os.Getenv("NVIDIA_DRIVER_INSTALLER_DAEMONSET")
+	if dsYamlUrlFromEnv != "" {
+		dsYamlUrl = dsYamlUrlFromEnv
 	} else {
-		dsYamlUrlFromEnv := os.Getenv("NVIDIA_DRIVER_INSTALLER_DAEMONSET")
-		if dsYamlUrlFromEnv != "" {
-			dsYamlUrl = dsYamlUrlFromEnv
-		} else {
-			dsYamlUrl = "https://raw.githubusercontent.com/GoogleCloudPlatform/container-engine-accelerators/master/daemonset.yaml"
-		}
-		gpuResourceName = framework.NVIDIAGPUResourceName
-		podCreationFunc = makeCudaAdditionDevicePluginTestPod
+		dsYamlUrl = "https://raw.githubusercontent.com/GoogleCloudPlatform/container-engine-accelerators/master/daemonset.yaml"
 	}
+	gpuResourceName = framework.NVIDIAGPUResourceName
 
 	framework.Logf("Using %v", dsYamlUrl)
 	// Creates the DaemonSet that installs Nvidia Drivers.
@@ -218,7 +168,7 @@ func testNvidiaGPUsOnCOS(f *framework.Framework) {
 	framework.Logf("Creating as many pods as there are Nvidia GPUs and have the pods run a CUDA app")
 	podList := []*v1.Pod{}
 	for i := int64(0); i < getGPUsAvailable(f); i++ {
-		podList = append(podList, f.PodClient().Create(podCreationFunc()))
+		podList = append(podList, f.PodClient().Create(makeCudaAdditionDevicePluginTestPod()))
 	}
 	framework.Logf("Wait for all test pods to succeed")
 	// Wait for all pods to succeed
@@ -233,13 +183,6 @@ func testNvidiaGPUsOnCOS(f *framework.Framework) {
 	f.TestSummaries = append(f.TestSummaries, summary)
 	framework.ExpectNoError(err, "getting resource usage summary")
 }
-
-var _ = SIGDescribe("[Feature:GPU]", func() {
-	f := framework.NewDefaultFramework("gpus")
-	It("run Nvidia GPU tests on Container Optimized OS only", func() {
-		testNvidiaGPUsOnCOS(f)
-	})
-})
 
 var _ = SIGDescribe("[Feature:GPUDevicePlugin]", func() {
 	f := framework.NewDefaultFramework("device-plugin-gpus")
